output/test-agent/requirements.txt
agentscope>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0